Moved Telemetry SimState to internal
