+++ SimTelemetry.Data/SimTelemetry.Data.csproj
@@ -80,7 +80,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Telemetry.cs" />
     <Compile Include="Simulators.cs" />
-    <Compile Include="Telemetry_SimState.cs" />
     <Compile Include="TrackMap.cs">
       <SubType>UserControl</SubType>
     </Compile>
