--- SimTelemetry.Data/Telemetry_SimState.cs
@@ -1,40 +1,0 @@
-﻿/*************************************************************************
- *                         SimTelemetry                                  *
- *        providing live telemetry read-out for simulators               *
- *             Copyright (C) 2011-2012 Hans de Jong                      *
- *                                                                       *
- *  This program is free software: you can redistribute it and/or modify *
- *  it under the terms of the GNU General Public License as published by *
- *  the Free Software Foundation, either version 3 of the License, or    *
- *  (at your option) any later version.                                  *
- *                                                                       *
- *  This program is distributed in the hope that it will be useful,      *
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of       *
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        *
- *  GNU General Public License for more details.                         *
- *                                                                       *
- *  You should have received a copy of the GNU General Public License    *
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.*
- *                                                                       *
- * Source code only available at https://github.com/nlhans/SimTelemetry/ *
- ************************************************************************/
-namespace SimTelemetry.Data
-{
-    /// <summary>
-    /// Class containing whether a simulator is running and has a session active.
-    /// </summary>
-    public class Telemetry_SimState
-    {
-        public bool Active;
-        public bool Session;
-        public bool Driving;
-        public int Laps;
-        public Telemetry_SimState()
-        {
-            Laps = -1;
-            Driving = false;
-            Active = false;
-            Session = false;
-        }
-    }
-}
